requirements.txt
Django==5.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
psycopg2-binary==2.9.9
redis==5.0.1
drf-spectacular==0.27.1
python-decouple==3.8
requests==2.31.0
daphne==4.1.0